separate StashCaption
