+++ src/content.js
@@ -3,6 +3,7 @@
 import React, { useState, useEffect } from "react";
 import ReactDOM from "react-dom";
 import ReactModal from "react-modal";
+import StashCaption from "./components/StashCaption"
 import StashEntry from "./components/StashEntry"
 import RestoreButton from "./components/RestoreButton"
 import "./css/content.css";
@@ -261,27 +262,6 @@ function StashEntries({stashKey, hours, minutes, entries, chromePort}) {
         if (chromePort != null)
           chromePort.postMessage(popStashEntry(stashKey))
       }} count={count} />
-    </div>
-  );
-}
-
-function StashCaption({hours, minutes, count}) {
-  function displayTime(hours, minutes) {
-    const q = Math.floor(hours / 12);
-    const r = hours % 12;
-    const mm = `${minutes}`.padStart(2, "0");
-    if (q < 1) return `${hours}:${mm} AM`;
-    else if (r == 0) return `12:${mm} PM`;
-    else return `${r}:${mm} PM`;
-  }
-  return (
-    <div className={"padding-vertical-medium"}>
-      <div class="shade-087 font-size-larger font-weight-bold line-height-medium">
-        {displayTime(hours, minutes)}
-      </div>
-      <div class="shade-056 font-size-small font-weight-medium line-height-medium">
-        {count} items
-      </div>
     </div>
   );
 }
